Fix bug with previous commit that caused FTBFS packages on non-Jaunty.
I previously misinterpreted the description of the --install-layout=deb
option. It turns out one cannot use this option with older Ubuntu versions.

This patch reverts to the old behavior for Python 2.4 and 2.5 but
otherwise uses the new behavior. Ideally, the test would actually
test whether the --install-layout=deb option exists, but this works
for now.

Still need to test on debian -- I don't know if their Python 2.6 has this
option.