don't import pkg_resources unless needed
pkg_resources may not be installed, and there's no need for it in many cases.
This fixes "python setup.py --command-packages=stdeb.command sdist_dsc" when
pkg_resources is not installed. In that case, stdeb.util silently failed to
import, causing distutils to claim that no sdist_dsc command was available.