always override dh_auto_install
I was running into a case where building only a Python 3 package using
python 2 on Ubuntu 12.04 ("--with-python3 true --with-python2 false")
was failing because there was no override triggered for
dh_auto_install and thus python (not python3) was called to install
the binary package. This fixes the problem by always overriding
dh_auto_install.