Use standard defaults formatting with special case.
In order to specify a default using the conventional format we must also
add a special case to the default regexp parser.
As noted we set the value to an empty string and the consuming code will
check for the empty string and use the value of the suite option
instead.