Add an Earthfile for use with earthly.
<https://earthly.dev> is an open source (MPL-2.0) project for defining
and executing builds using containers.

It meets a lot of the same needs as the Vagrant boxes that are in this
project using containers rather than VMs.

The advantage of pairing this with GitHub Actions is that it allows
maintainers and contributors to reproduce test results locally which is
very difficult when using GitHub Actions natively.

stdeb interacts with a lot of system-level utilities so having
convenient testing across multiple supported distributions is highly
valuable.