Switch to a combination of the "Simple" and JSON APIs for PyPI.
The XML-RPC API is deprecated. Tests using it were failing due to what
seems like a malformed url. I have no idea if this API has already been
dropped or if there's some usage issue inside stdeb but I was not
motivated to investigate since the entire API is deprecated.

Instead I replaced the functionality with a new Requests HTTP session
client that uses the [JSON representation][1] of the [Simple API][2] and
the [JSON API][3] where best appropriate. It could all be done with the
JSON API but parts of it are also deprecated...

There is one regression in behavior, which is that packages with a
release on PyPI but no download url on PyPI will not be checked for an
external download url as this also deprecated.

[1]: https://peps.python.org/pep-0691/
[2]: https://warehouse.pypa.io/api-reference/legacy.html
[3]: https://warehouse.pypa.io/api-reference/json.html