Check path for python2 binary name when generating deb info.
Ubuntu Focal and Jammy install python2 with python-all-dev.
Rather than requiring that tests and downstream packages always install
python-is-python2 (seemingly unavailable on jammy), check the current
path for each binary and use the first one found, preferring `python`
when it is available to maintain current behavior but using `python2` if
it's the only option.

This unblocks testing python2 package generation with python3-stdeb in
test-2and3.sh.